Move assign group request out to it's own file.
