+++ app/com/gu/identity/frontend/services/IdentityService.scala
@@ -5,7 +5,7 @@ import com.gu.identity.frontend.controllers.RegisterRequest
 import com.gu.identity.frontend.models.{ClientRegistrationIp, TrackingData}
 import com.gu.identity.service.client.models.User
 import com.gu.identity.service.client._
-import com.gu.identity.service.client.request.RegisterApiRequest
+import com.gu.identity.service.client.request.{AssignGroupApiRequest, RegisterApiRequest}
 import org.joda.time.{DateTime, Seconds}
 import play.api.mvc.{Cookie => PlayCookie}
 
@@ -93,7 +93,7 @@ class IdentityServiceImpl(config: Configuration, adapter: IdentityServiceRequest
   }
 
   override def assignGroupCode(group: String, cookie: PlayCookie)(implicit ec: ExecutionContext): Future[Either[Seq[ServiceError], AssignGroupResponse]] = {
-    val apiRequest = AssignGroupRequest(group, cookie)
+    val apiRequest = AssignGroupApiRequest(group, cookie)
     client.assignGroupCode(apiRequest).map {
       case Left(errors) => Left {
         errors.map {
